+++ AdapterUnitTests/UnityAdapterTests/AUTUnityAdapterTests.m
@@ -243,4 +243,124 @@
   [self waitForExpectations:@[ expectation ]];
 }
 
+#pragma mark - Privacy Preferences Tests
+
+- (void)assertPrivacyPreferencesWithChildDirected:(nullable NSNumber *)childDirected
+                                         underAge:(nullable NSNumber *)underAge
+                           ageRestrictedTreatment:(GADAgeRestrictedTreatment)ageRestrictedTreatment
+                            expectedNonBehavioral:(BOOL)expectedNonBehavioral {
+  GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = childDirected;
+  GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = underAge;
+  GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment = ageRestrictedTreatment;
+
+  id metaDataMock = OCMClassMock([UADSMetaData class]);
+  OCMStub([metaDataMock alloc]).andReturn(metaDataMock);
+  OCMExpect([metaDataMock set:@"user.nonbehavioral" value:@(expectedNonBehavioral)]);
+  OCMExpect([metaDataMock commit]);
+
+  [GADMediationAdapterUnity updatePrivacyPreferences];
+
+  OCMVerifyAll(metaDataMock);
+  [metaDataMock stopMocking];
+}
+
+- (void)testUpdatePrivacyPreferences_childDirectedTrueAndUnderAgeTrue_setsNonBehavioralToYes {
+  [self assertPrivacyPreferencesWithChildDirected:@YES
+                                         underAge:@YES
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentUnspecified
+                            expectedNonBehavioral:YES];
+}
+
+- (void)testUpdatePrivacyPreferences_childDirectedTrueAndUnderAgeFalse_setsNonBehavioralToYes {
+  [self assertPrivacyPreferencesWithChildDirected:@YES
+                                         underAge:@NO
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentUnspecified
+                            expectedNonBehavioral:YES];
+}
+
+- (void)testUpdatePrivacyPreferences_childDirectedTrue_setsNonBehavioralToYes {
+  [self assertPrivacyPreferencesWithChildDirected:@YES
+                                         underAge:nil
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentUnspecified
+                            expectedNonBehavioral:YES];
+}
+
+- (void)testUpdatePrivacyPreferences_childDirectedFalseAndUnderAgeTrue_setsNonBehavioralToYes {
+  [self assertPrivacyPreferencesWithChildDirected:@NO
+                                         underAge:@YES
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentUnspecified
+                            expectedNonBehavioral:YES];
+}
+
+- (void)testUpdatePrivacyPreferences_underAgeTrue_setsNonBehavioralToYes {
+  [self assertPrivacyPreferencesWithChildDirected:nil
+                                         underAge:@YES
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentUnspecified
+                            expectedNonBehavioral:YES];
+}
+
+- (void)testUpdatePrivacyPreferences_childDirectedFalseAndUnderAgeFalse_setsNonBehavioralToNo {
+  [self assertPrivacyPreferencesWithChildDirected:@NO
+                                         underAge:@NO
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentUnspecified
+                            expectedNonBehavioral:NO];
+}
+
+- (void)testUpdatePrivacyPreferences_childDirectedFalseAndUnderAgeUnspecified_setsNonBehavioralToNo {
+  [self assertPrivacyPreferencesWithChildDirected:@NO
+                                         underAge:nil
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentUnspecified
+                            expectedNonBehavioral:NO];
+}
+
+- (void)testUpdatePrivacyPreferences_childDirectedUnspecifiedAndUnderAgeFalse_setsNonBehavioralToNo {
+  [self assertPrivacyPreferencesWithChildDirected:nil
+                                         underAge:@NO
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentUnspecified
+                            expectedNonBehavioral:NO];
+}
+
+- (void)testUpdatePrivacyPreferences_ageRestrictedChild_setsNonBehavioralToYes {
+  [self assertPrivacyPreferencesWithChildDirected:nil
+                                         underAge:nil
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentChild
+                            expectedNonBehavioral:YES];
+}
+
+- (void)
+    testUpdatePrivacyPreferences_legacyUnspecifiedOverridesAgeRestrictedTeen_setsNonBehavioralToNo {
+  [self assertPrivacyPreferencesWithChildDirected:nil
+                                         underAge:nil
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentTeen
+                            expectedNonBehavioral:NO];
+}
+
+- (void)testUpdatePrivacyPreferences_allUnspecified_setsNonBehavioralToNo {
+  [self assertPrivacyPreferencesWithChildDirected:nil
+                                         underAge:nil
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentUnspecified
+                            expectedNonBehavioral:YES];
+}
+
+- (void)testUpdatePrivacyPreferences_ageRestrictedChildOverridesLegacyAdult_setsNonBehavioralToYes {
+  [self assertPrivacyPreferencesWithChildDirected:@NO
+                                         underAge:@NO
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentChild
+                            expectedNonBehavioral:YES];
+}
+
+- (void)testUpdatePrivacyPreferences_ageRestrictedTeenWithLegacyAdult_setsNonBehavioralToNo {
+  [self assertPrivacyPreferencesWithChildDirected:@NO
+                                         underAge:@NO
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentTeen
+                            expectedNonBehavioral:NO];
+}
+
+- (void)testUpdatePrivacyPreferences_legacyChildOverridesAgeRestrictedTeen_setsNonBehavioralToYes {
+  [self assertPrivacyPreferencesWithChildDirected:@YES
+                                         underAge:nil
+                           ageRestrictedTreatment:GADAgeRestrictedTreatmentTeen
+                            expectedNonBehavioral:YES];
+}
+
 @end