+++ adapters/Unity/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Next version
 - Adapter now automatically forwards GDPR consent found inside `IABTCF_AddtlConsent` to the Unity Ads SDK if GDPR applies.
+- Maps `ageRestrictedTreatment` to Unity Ads SDK's privacy settings.
 
 #### [Version 4.19.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/unity/UnityAdapter-4.19.0.0.zip)
 - Verified compatibility with Unity Ads SDK 4.19.0.